requirements.txt
torch>=2.1.0
transformers>=4.35.0
accelerate>=0.25.0
ray[rllib]>=2.8.0
anthropic>=0.8.0
openai>=1.6.0
google-generativeai>=0.3.0
bandit>=1.7.5
semgrep>=1.50.0
fastapi>=0.108.0
uvicorn>=0.25.0
docker>=7.0.0
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0
pytest>=7.4.3
pyyaml>=6.0.1
python-dotenv>=1.0.0
rich>=13.7.0
tqdm>=4.66.1
matplotlib>=3.8.2
tenacity>=8.2.3